+++ CHANGES.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Improve end turn warning to be less aggressive
   * If you cannot attack hexagons by buying new units the warning will not be displayed (unless you can buy and place a castle)
 * Capitals behave like living pieces and will not instantly become a tree, but rather be replaced with a tree at the start of the owning teams turn
+* Only grow trees once a round
 
 ### Fixed